Initial pass at all methods and tests done.
I think I have full coverage of all XMLStreamWriter methods,
except for support for NamespaceContext, which I'm not supporting
until a use case need for it arises.

All methods have tests except for a few of the simplest pass throughs
like writeDTD().

The javadoc needs work and the maven dir structure is not wired up
correctly, so still need to fix those.